HealthKit integration (#11)
This PR integrate bridge-ios-healthkit in the application and synchronizes the events in Pryv with the samples in HealthKit:

If authorized, the application will get the updates from HealthKit for date of birth, wheelchair use, body mass, height, body mass index, active energy burned and workout. These updates will trigger the creation of a new event in Pryv.io backend with the content of the sample received from HealthKit.

The application will also create a HealthKit sample for body mass upon the creation of a simple event with stream id bodyMass in the application.